✓ win9x: property sheet
git-svn-id: http://amethyst.yui.ne.jp/svn/pc98/np2/branches/yui@2479 7c2116fe-53d0-4aec-ab59-80607f05abc0

+++ reform_ui/win9x/misc/DlgProc.h
@@ -98,7 +98,7 @@ public:
 	 */
 	int FindString(int nStartAfter, LPCTSTR lpszString) const
 	{
-		return ::SendMessage(m_hWnd, CB_FINDSTRING, nStartAfter, reinterpret_cast<LPARAM>(lpszString));
+		return static_cast<int>(::SendMessage(m_hWnd, CB_FINDSTRING, nStartAfter, reinterpret_cast<LPARAM>(lpszString)));
 	}
 
 	/**
@@ -109,7 +109,7 @@ public:
 	 */
 	int FindStringExact(int nStartAfter, LPCTSTR lpszString) const
 	{
-		return ::SendMessage(m_hWnd, CB_FINDSTRINGEXACT, nStartAfter, reinterpret_cast<LPARAM>(lpszString));
+		return static_cast<int>(::SendMessage(m_hWnd, CB_FINDSTRINGEXACT, nStartAfter, reinterpret_cast<LPARAM>(lpszString)));
 	}
 
 	/**
